pyproject.toml
[project]
name = "lox"
version = "0.1.0"
description = "Prova 2 de Compiladores - Interpretação de Lox"
readme = "README.md"
authors = [
    { name = "Fábio Macêdo Mendes", email = "fabiomacedomendes@gmail.com" },
]
requires-python = ">=3.10"
dependencies = ["ipdb>=0.13.13", "lark-parser>=0.12.0", "rich>=14.0.0"]

[project.scripts]
lox = "lox.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.3.5"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["F403", "F405"]
